adjacency matrix + list comparison

+++ src/com/inf/unibz/entity/Edge.java
@@ -6,9 +6,9 @@ public class Edge  {
 	  private final int id; 
 	  private final Vertex source;
 	  private final Vertex destination;
-	  private final int weight; 
+	  private final double weight; 
 	  
-	  public Edge(int id, Vertex source, Vertex destination, int weight) {
+	  public Edge(int id, Vertex source, Vertex destination, double weight) {
 	    this.id = id;
 	    this.source = source;
 	    this.destination = destination;
@@ -25,7 +25,7 @@ public class Edge  {
 	  public Vertex getSource() {
 	    return source;
 	  }
-	  public int getWeight() {
+	  public double getWeight() {
 	    return weight;
 	  }
